% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVDML-package.R
\docType{package}
\name{IVDML-package}
\alias{IVDML}
\alias{IVDML-package}
\title{IVDML: Double Machine Learning with Instrumental Variables and Heterogeneity}
\description{
Instrumental variable (IV) estimators for homogeneous and heterogeneous treatment effects with efficient machine learning instruments. The estimators are based on double/debiased machine learning allowing for nonlinear and potentially high-dimensional control variables. Details can be found in Scheidegger, Guo and Bühlmann (2025) "Inference for heterogeneous treatment effects with efficient instruments and machine learning" \href{https://arxiv.org/abs/2503.03530}{arXiv:2503.03530}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cyrillsch/IVDML}
}

}
\author{
\strong{Maintainer}: Cyrill Scheidegger \email{cyrill.scheidegger@stat.math.ethz.ch} (\href{https://orcid.org/0009-0005-2851-1384}{ORCID}) [copyright holder]

}
\keyword{internal}
