% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVDML-package.R
\docType{package}
\name{IVDML-package}
\alias{IVDML}
\alias{IVDML-package}
\title{IVDML: Double Machine Learning for potentially heterogeneous treatment effects with Instrumental Variables}
\description{
Instrumental variable (IV) estimators for homogeneous and heterogeneous treatment effects with efficient machine learning instruments. The estimators are based on double/debiased machine learning allowing for nonlinear and potentially high-dimensional control variables. Details can be found in ?Scheidegger et al. 2025 (to be published)?.
}
\author{
\strong{Maintainer}: Cyrill Scheidegger \email{cyrill.scheidegger@stat.math.ethz.ch}

}
\keyword{internal}
